% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machinery.R
\name{machineryRegisterTask}
\alias{machineryRegisterTask}
\title{Registers a task handler for the given taskName.}
\usage{
machineryRegisterTask(machinery, taskName, callback)
}
\arguments{
\item{machinery}{The machinery instance to use}

\item{taskName}{The name of the task that the handler handles}

\item{callback}{The callback function to invoke when machinery gets this task}
}
\value{
The modified machinery instance
}
\description{
Registers a task handler for the given taskName.
}
