% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machinery.R
\name{machinerySendSimpleTask}
\alias{machinerySendSimpleTask}
\title{Sends a simple task to the given queue.}
\usage{
machinerySendSimpleTask(machinery, queue, taskName, args = list())
}
\arguments{
\item{machinery}{The machinery instance to use}

\item{queue}{The target queue to send the task into}

\item{taskName}{The name of the task}

\item{args}{Argument given to the task. Data frame with column "Type" and "Value"}
}
\value{
The created and sent task
}
\description{
Sends a simple task to the given queue.
}
