% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machinery.R
\name{machineryStart}
\alias{machineryStart}
\title{Start consuming tasks for the given machinery instance.}
\usage{
machineryStart(machinery, pollSecs = DEFAULT_POLL_PERIOD)
}
\arguments{
\item{machinery}{The machinery instance to use.}

\item{pollSecs}{The number of seconds to wait before polling again for new tasks.}
}
\value{
None
}
\description{
This function is a blocking function.
}
