% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machinery.R
\name{machineryWaitForResults}
\alias{machineryWaitForResults}
\title{Waits for result of the given task indefinitely.}
\usage{
machineryWaitForResults(machinery, task, pollSecs = DEFAULT_POLL_PERIOD)
}
\arguments{
\item{machinery}{The machinery instance to use}

\item{task}{The machinery task to get the results from}

\item{pollSecs}{The number of seconds to rest before polling again}
}
\value{
A data frame containing the results, with column "Type" and "Value"
}
\description{
This function is a blocking function.
}
